+++ drizzle/20260919132009_adorable_pete_wisdom/migration.sql
@@ -1,0 +1,2 @@
+ALTER TABLE "links" ADD COLUMN "limit_destination" text;--> statement-breakpoint
+ALTER TABLE "links" ADD COLUMN "scheduled_destination" text;